// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/phrase-context-ayah-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PhraseAyahWordDto } from '../models/phrase-ayah-word-dto';
import { PhraseContextHighlightsDto } from '../models/phrase-context-highlights-dto';
export interface PhraseContextAyahDto {
  ayahId: number;
  ayahNumber: number;
  highlights: PhraseContextHighlightsDto;
  occurrenceCount: number;
  pageFrom: number;
  pageTo: number;
  surahNameArabic: string;
  surahNumber: number;
  verseKey: string;
  words: Array<PhraseAyahWordDto>;
}